requirements.txt
torch
spaces
gradio
transformers==4.50.2
accelerate
pytest
loguru
python-dotenv
opencv-python
timm